data-generator/requirements.txt
kafka-python==2.0.2
pandas==2.0.1
numpy==1.24.3
faker==18.9.0
python-dotenv==1.0.0